+++ client/Packages/com.beamable/Runtime/Player/CloudSaving/Data/ConflictResolveType.cs
@@ -1,0 +1,8 @@
+﻿namespace Beamable.Player.CloudSaving
+{
+	public enum ConflictResolveType
+	{
+		UseLocal,
+		UseCloud,
+	}
+}